requirements.txt
matplotlib==3.4.3
mlflow==2.10.2
numpy==1.24.4
opencv_python==4.5.4.60
PyYAML==6.0.1
PyYAML==6.0.1
torch==1.10.0
torchvision==0.11.1
